# Style Refresh: put stylesheets into single style tags