This PR is an attempt to put every stylesheet in its dedicated `<style>` tag in the document's header.
It's a very simple approach here and it works really well.

Look at the [TiddlyFlex Layout](https://burningtreec.github.io/TiddlyFlex/) where this is enabled for testing purposes and the performance instrumentation is enabled.
Drag the sidebar resizer around and notice that the styleRefresh is very low.
In the TiddlyFlex Layout this is achieved by a "refresh-blocker" widget that blocks refreshing down the widget-tree if a given tiddler changes and another given tiddler exists.
The UI is wrapped within refresh-blocker widgets which makes the mainRefresh 0-1ms when dragging the sidebar.
The stylesheets that don't need to be refreshes when the tiddlers change that make the sidebar resize are also wrapped within refresh-blocker widgets. That makes the styleRefresh 3-5ms when dragging the sidebar.

The idea would be to add a refresh-blocker to the core, also add a sidebar resizer and make use of the upsides of this.

What do you think?