@pmario 2024-09-22
hmmm. Some very short testing with TW-com and TiddlyFlex from your link.

- With TW-com the styleRefresh typically is 4ms to 6ms. Worst case 10ms on my PC

- With your TiddlyFlex wiki it is typically 10ms to 14ms. Wors case 23ms

So there is a severe performance regression going on. 

I did test it with **right sidebar -> More -> * tabs** .. Switching between all vertical tabs. 

The problem is, that styleRefresh is triggered with every keystroke in edit-mode. -- IMO If we are able to fix that for tiddlywiki.com, I thin as a side effect it will also resolve the problem you have with TiddlyFlex.